+++ src/App.tsx
@@ -25,7 +25,7 @@ import '@ionic/react/css/display.css';
 
 /* Theme variables */
 import './theme/variables.css';
-import Home from './components/Home';
+import Home from './Containers/Home';
 import Favourites from './Containers/Favourites';
 import Tech from './Containers/Tech';
